Extract path_range.rs
